refactor/#25 : User Entity 로 폴더 변경
